+++ admin/src/components/UserCard.jsx
@@ -1,6 +1,9 @@
 import React from "react";
-import { ROLES, translate } from "snu-lib";
+import { ROLES } from "snu-lib";
 import cx from "classnames";
+
+import UserInfo from "./UserInfo";
+import LinkWrapper from "./LinkWrapper";
 
 export default function UserCard({ user }) {
   function getAvatar(user) {
@@ -16,72 +19,36 @@ export default function UserCard({ user }) {
     if (user?.role === "Volontaire") return `/volontaire/${user._id}`;
     return null;
   }
-  function getRole(user) {
-    if (!user) return "Donnée indisponible";
-    if (user.role) {
-      return translate(user.role);
-    } else {
-      return "Script";
-    }
-  }
-  function getAuthor(user) {
-    if (!user) return "Auteur inconnu";
-    if (user.firstName && user.lastName) return `${user.firstName} ${user.lastName.toUpperCase()}`;
-    if (user.firstName && !user.lastName) return user.firstName;
-  }
+  const hasLink = getLink(user) !== null;
 
   return (
     <div className="flex items-center gap-2">
       <div
         className={cx("flex h-10 w-10 shrink-0 items-center justify-center rounded-full border-2 border-white bg-slate-100 font-medium uppercase ", {
-          "text-blue-600 transition group-hover:bg-blue-600 group-hover:text-slate-100": getLink(user) !== null,
-          "text-cyan-900": getLink(user) === null,
+          "text-blue-600 transition group-hover:bg-blue-600 group-hover:text-slate-100": hasLink,
+          "text-cyan-900": !hasLink,
         })}>
         {getAvatar(user)}
       </div>
-      {user.impersonatedBy ? (
-        <div className="flex w-10/12 flex-col leading-5">
-          <a
-            href={getLink(user)}
-            target="_blank"
-            rel="noreferrer"
-            className={cx("group flex w-full flex-col", {
-              "hover:cursor-pointer hover:text-blue-600": getLink(user) !== null,
-              "hover:text-inherit": getLink(user) === null,
-            })}>
-            <p className="w-full truncate font-medium decoration-2 underline-offset-2 text-sm">
-              {getAuthor(user.impersonatedBy)} <span className="truncate font-normal capitalize text-gray-500 text-xs">{getRole(user)}</span>
-            </p>
-          </a>
-          <a
-            href={getLink(user)}
-            target="_blank"
-            rel="noreferrer"
-            className={cx("group flex w-full flex-col", {
-              "hover:cursor-pointer hover:text-blue-600": getLink(user) !== null,
-              "hover:text-inherit": getLink(user) === null,
-            })}>
-            <p className="w-full truncate capitalize text-gray-500 text-xs decoration-2 underline-offset-2">En tant que :</p>
-            <p className="w-full truncate font-medium decoration-2 underline-offset-2 text-sm">
-              {getAuthor(user)} <span className="truncate font-normal capitalize text-gray-500 text-xs">{getRole(user)}</span>
-            </p>
-          </a>
-        </div>
-      ) : (
-        <a
-          href={getLink(user)}
-          target="_blank"
-          rel="noreferrer"
-          className={cx("group flex w-full flex-col", {
-            "hover:cursor-pointer hover:text-blue-600": getLink(user) !== null,
-            "hover:text-inherit": getLink(user) === null,
-          })}>
-          <div className="flex w-10/12 flex-col leading-5">
-            <p className="w-full truncate font-medium decoration-2 underline-offset-2 text-sm">{getAuthor(user)}</p>
-            <p className="w-full truncate capitalize text-gray-500 text-xs decoration-2 underline-offset-2">{getRole(user)}</p>
-          </div>
-        </a>
-      )}
+
+      <div className="flex w-10/12 flex-col leading-5">
+        {user.impersonatedBy ? (
+          <>
+            <LinkWrapper link={getLink(user.impersonatedBy)} hasLink={getLink(user.impersonatedBy) !== null}>
+              <UserInfo user={user.impersonatedBy} />
+            </LinkWrapper>
+
+            <LinkWrapper link={getLink(user)} hasLink={hasLink}>
+              <p className="w-full truncate capitalize text-gray-500 text-xs decoration-2 underline-offset-2">En tant que :</p>
+              <UserInfo user={user} />
+            </LinkWrapper>
+          </>
+        ) : (
+          <LinkWrapper link={getLink(user)} hasLink={hasLink}>
+            <UserInfo user={user} />
+          </LinkWrapper>
+        )}
+      </div>
     </div>
   );
 }
